enterprise-knowledge-chatbot-rag/requirements.txt
fastapi
uvicorn[standard]
requests
rapidfuzz
python-telegram-bot==20.7
python-dotenv
langchain
langchain-openai
faiss-cpu
python-dotenv
streamlit
sentence-transformers
torch
pypdf
numpy